[WIP] Extract QP download code from BindingProcessImpl
