+++ src/ConnectedMode.UnitTests/Binding/BindingProcessImplTests.cs
@@ -19,20 +19,15 @@
  */
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using SonarLint.VisualStudio.ConnectedMode.Binding;
+using SonarLint.VisualStudio.ConnectedMode.QualityProfiles;
 using SonarLint.VisualStudio.Core.Analysis;
-using SonarLint.VisualStudio.Core.Binding;
 using SonarLint.VisualStudio.Integration;
-using SonarLint.VisualStudio.Integration.Resources;
-using SonarLint.VisualStudio.Progress.Controller;
 using SonarLint.VisualStudio.TestInfrastructure;
 using SonarQube.Client;
 using SonarQube.Client.Models;
-using Language = SonarLint.VisualStudio.Core.Language;
 
 namespace SonarLint.VisualStudio.ConnectedMode.UnitTests.Binding
 {
@@ -40,46 +35,6 @@ namespace SonarLint.VisualStudio.ConnectedMode.UnitTests.Binding
     public class BindingProcessImplTests
     {
         #region Tests
-
-        [TestMethod]
-        public void Ctor_ArgChecks()
-        {
-            var bindingArgs = CreateBindCommandArgs(connection: new ConnectionInformation(new Uri("http://server")));
-            var slnBindOp = Mock.Of<ISolutionBindingOperation>();
-            var bindingConfigProvider = Mock.Of<IBindingConfigProvider>();
-            var sonarQubeService = Mock.Of<ISonarQubeService>();
-            var configurationPersister = Mock.Of<IConfigurationPersister>();
-            var exclusionSettingsStorage = Mock.Of<IExclusionSettingsStorage>();
-            var logger = Mock.Of<ILogger>();
-
-            // 1. Null binding args
-            Action act = () => new BindingProcessImpl(null, slnBindOp, bindingConfigProvider, exclusionSettingsStorage, configurationPersister, sonarQubeService, logger);
-            act.Should().ThrowExactly<ArgumentNullException>().And.ParamName.Should().Be("bindingArgs");
-
-            // 2. Null solution binding operation
-            act = () => new BindingProcessImpl(bindingArgs, null, bindingConfigProvider, exclusionSettingsStorage, configurationPersister, sonarQubeService, logger);
-            act.Should().ThrowExactly<ArgumentNullException>().And.ParamName.Should().Be("solutionBindingOperation");
-
-            // 3. Null binding configuration provider
-            act = () => new BindingProcessImpl(bindingArgs, slnBindOp, null, exclusionSettingsStorage, configurationPersister, sonarQubeService, logger);
-            act.Should().ThrowExactly<ArgumentNullException>().And.ParamName.Should().Be("bindingConfigProvider");
-
-            // 4. Null exclusion settings storage
-            act = () => new BindingProcessImpl(bindingArgs, slnBindOp, bindingConfigProvider, null, configurationPersister, sonarQubeService, logger);
-            act.Should().ThrowExactly<ArgumentNullException>().And.ParamName.Should().Be("exclusionSettingsStorage");
-
-            // 5. Null configuration persister
-            act = () => new BindingProcessImpl(bindingArgs, slnBindOp, bindingConfigProvider, exclusionSettingsStorage, null, sonarQubeService, logger);
-            act.Should().ThrowExactly<ArgumentNullException>().And.ParamName.Should().Be("configurationPersister");
-
-            // 6. Null SonarQubeService
-            act = () => new BindingProcessImpl(bindingArgs, slnBindOp, bindingConfigProvider, exclusionSettingsStorage, configurationPersister, null, logger);
-            act.Should().ThrowExactly<ArgumentNullException>().And.ParamName.Should().Be("sonarQubeService");
-
-            // 7. Null logger
-            act = () => new BindingProcessImpl(bindingArgs, slnBindOp, bindingConfigProvider, exclusionSettingsStorage, configurationPersister, sonarQubeService, null);
-            act.Should().ThrowExactly<ArgumentNullException>().And.ParamName.Should().Be("logger");
-        }
 
         [TestMethod]
         public async Task SaveServerExclusionsAsync_ReturnsTrue()
@@ -156,190 +111,20 @@ namespace SonarLint.VisualStudio.ConnectedMode.UnitTests.Binding
         }
 
         [TestMethod]
-        public async Task DownloadQualityProfile_SavesConfiguration()
+        public async Task DownloadQualityProfile_CreatesBoundProjectAndCallsQPDownloader()
         {
-            var configPersister = new DummyConfigPersister();
+            // TODO duncanp
 
-            // Arrange
-            const string QualityProfileName = "SQQualityProfileName";
-            const string ProjectName = "SQProjectName";
+            //var bindingArgs = CreateBindCommandArgs("key", ProjectName, new ConnectionInformation(new Uri("http://connected")));
+            //var testSubject = CreateTestSubject(bindingArgs,
+            //    configProviderMock.Object,
+            //    sonarQubeService: sonarQubeService.Object,
+            //    configurationPersister: configPersister,
+            //    languagesToBind: new []{ language });
 
-            var progressAdapter = Mock.Of<IProgress<FixedStepsProgress>>();
+            //// Act
+            //await testSubject.DownloadQualityProfileAsync(progressAdapter, CancellationToken.None);
 
-            var bindingConfig = new Mock<IBindingConfig>().Object;
-
-            var language = Language.VBNET;
-            var sonarQubeService = new Mock<ISonarQubeService>();
-            SonarQubeQualityProfile profile = ConfigureQualityProfile(sonarQubeService, language, QualityProfileName);
-
-            var configProviderMock = new Mock<IBindingConfigProvider>();
-            configProviderMock.Setup(x => x.GetConfigurationAsync(profile, language, It.IsAny<BindingConfiguration>(), CancellationToken.None))
-                .ReturnsAsync(bindingConfig);
-
-            var bindingArgs = CreateBindCommandArgs("key", ProjectName, new ConnectionInformation(new Uri("http://connected")));
-            var testSubject = CreateTestSubject(bindingArgs,
-                configProviderMock.Object,
-                sonarQubeService: sonarQubeService.Object,
-                configurationPersister: configPersister,
-                languagesToBind: new []{ language });
-
-            // Act
-            await testSubject.DownloadQualityProfileAsync(progressAdapter, CancellationToken.None);
-
-            // Assert
-            configPersister.SavedProject.Should().NotBeNull();
-
-            var savedProject = configPersister.SavedProject;
-            savedProject.ServerUri.Should().Be(bindingArgs.Connection.ServerUri);
-            savedProject.Profiles.Should().HaveCount(1);
-            savedProject.Profiles[Language.VBNET].ProfileKey.Should().Be(profile.Key);
-            savedProject.Profiles[Language.VBNET].ProfileTimestamp.Should().Be(profile.TimeStamp);
-        }
-
-        [TestMethod]
-        public async Task DownloadQualityProfile_WhenQualityProfileIsNotAvailable_OtherLanguagesDownloadedSucceessfully()
-        {
-            // Arrange
-            var logger = new TestLogger(logToConsole: true);
-            var languagesToBind = new[] {
-                Language.Cpp,       // unavailable
-                Language.CSharp,
-                Language.Secrets,   // unavailable
-                Language.VBNET
-            };
-
-            // Configure available languages on the server
-            var configProvider = new Mock<IBindingConfigProvider>();
-            var sonarQubeService = new Mock<ISonarQubeService>();
-            SetupAvailableLanguages(configProvider, sonarQubeService,
-                Language.CSharp,
-                Language.VBNET,
-                Language.Css /* available on server, but shouldn't be requested*/ );
-
-            var configPersister = new DummyConfigPersister();
-
-            var testSubject = CreateTestSubject(
-                bindingConfigProvider: configProvider.Object,
-                configurationPersister: configPersister,
-                languagesToBind: languagesToBind,
-                sonarQubeService: sonarQubeService.Object,
-                logger: logger);
-
-            var notifications = new ConfigurableProgressStepExecutionEvents();
-            var progressAdapter = new Mock<IProgress<FixedStepsProgress>>();
-            progressAdapter.Setup(x => x.Report(It.IsAny<FixedStepsProgress>()))
-                .Callback<FixedStepsProgress>(x => ((IProgressStepExecutionEvents)notifications).ProgressChanged(x.Message, x.CurrentStep));
-
-            // Act
-            var result = await testSubject.DownloadQualityProfileAsync(progressAdapter.Object, CancellationToken.None);
-
-            // Assert
-            result.Should().BeTrue();
-            testSubject.InternalState.BindingConfigs.Should().NotContainKey(Language.Cpp);
-            testSubject.InternalState.BindingConfigs.Should().NotContainKey(Language.Secrets);
-            testSubject.InternalState.BindingConfigs.Should().NotContainKey(Language.Css);
-
-            testSubject.InternalState.BindingConfigs.Should().ContainKey(Language.CSharp);
-            testSubject.InternalState.BindingConfigs.Should().ContainKey(Language.VBNET);
-
-            // Progess notifications - percentage complete and messages
-            notifications.AssertProgress(1.0, 2.0, 3.0, 4.0);
-            CheckProgressMessages(languagesToBind);
-
-            // Check output messages
-            var missingPluginMessageCFamily = string.Format(Strings.SubTextPaddingFormat,
-                string.Format(BindingStrings.CannotDownloadQualityProfileForLanguage, Language.Cpp.Name));
-            var missingPluginMessageSecrets = string.Format(Strings.SubTextPaddingFormat,
-                string.Format(BindingStrings.CannotDownloadQualityProfileForLanguage, Language.Secrets.Name));
-
-            logger.AssertOutputStringExists(missingPluginMessageCFamily);
-            logger.AssertOutputStringExists(missingPluginMessageSecrets);
-            logger.AssertPartialOutputStringDoesNotExist(Language.Css.Name);
-
-            void CheckProgressMessages(params Language[] languages)
-            {
-                var expected = languages.Select(GetDownloadProgressMessages).ToArray();
-                notifications.AssertProgressMessages(expected);
-            }
-
-            static string GetDownloadProgressMessages(Language language)
-                => string.Format(BindingStrings.DownloadingQualityProfileProgressMessage, language.Name);
-        }
-
-        private void SetupAvailableLanguages(Mock<IBindingConfigProvider> configProvider,
-            Mock<ISonarQubeService> sonarQubeService,
-            params Language[] languages)
-        {
-            foreach (var language in languages)
-            {
-                var profile = ConfigureQualityProfile(sonarQubeService, language, "Profile" + language.Name);
-                configProvider.Setup(x => x.GetConfigurationAsync(profile, language, It.IsAny<BindingConfiguration>(), CancellationToken.None))
-                    .ReturnsAsync(Mock.Of<IBindingConfig>());
-            }
-        }
-
-        [TestMethod]
-        public async Task DownloadQualityProfile_WhenBindingConfigIsNull_Fails()
-        {
-            // Arrange
-            var logger = new TestLogger(logToConsole: true);
-            var configPersister = new DummyConfigPersister();
-            const string QualityProfileName = "SQQualityProfileName";
-            const string ProjectName = "SQProjectName";
-
-           var progressAdapter = Mock.Of<IProgress<FixedStepsProgress>>();
-
-            var language = Language.VBNET;
-            var sonarQubeService = new Mock<ISonarQubeService>();
-            SonarQubeQualityProfile profile = ConfigureQualityProfile(sonarQubeService, language, QualityProfileName);
-
-            var configProviderMock = new Mock<IBindingConfigProvider>();
-            var bindingArgs = CreateBindCommandArgs("key", ProjectName, new ConnectionInformation(new Uri("http://connected")));
-            var testSubject = CreateTestSubject(bindingArgs, configProviderMock.Object,
-                configurationPersister: configPersister,
-                languagesToBind: new[] { language },
-                sonarQubeService: sonarQubeService.Object,
-                logger: logger);
-
-            // Act
-            var result = await testSubject.DownloadQualityProfileAsync(progressAdapter, CancellationToken.None);
-
-            // Assert
-            result.Should().BeFalse();
-            testSubject.InternalState.QualityProfiles[language].Should().Be(profile);
-
-            logger.AssertOutputStrings(1);
-            var expectedOutput = string.Format(Strings.SubTextPaddingFormat,
-                string.Format(BindingStrings.FailedToCreateBindingConfigForLanguage, language.Name));
-            logger.AssertOutputStrings(expectedOutput);
-        }
-
-        [TestMethod]
-        public void GetBindingLanguages_ReturnsExpectedLanguages()
-        {
-            // Arrange
-            var expectedLanguages = new[] { Language.CSharp, Language.VBNET };
-
-            var testSubject = this.CreateTestSubject(languagesToBind: expectedLanguages);
-
-            // Act
-            var actualLanguages = testSubject.GetBindingLanguages();
-
-            // Assert
-            CollectionAssert.AreEquivalent(expectedLanguages, actualLanguages.ToArray(), "Unexpected languages for binding projects");
-        }
-
-        [TestMethod]
-        public void GetBindingLanguages_PublicCtor_ReturnsAllKnownLanguages()
-        {
-            // Arrange
-            var testSubject = this.CreateTestSubject();
-
-            // Act
-            var actualLanguages = testSubject.GetBindingLanguages();
-
-            // Assert
-            CollectionAssert.AreEquivalent(Language.KnownLanguages.ToArray(), actualLanguages.ToArray(), "Unexpected languages for binding projects");
         }
 
         [TestMethod]
@@ -352,43 +137,28 @@ namespace SonarLint.VisualStudio.ConnectedMode.UnitTests.Binding
             testSubject.InternalState.BindingOperationSucceeded.Should().BeTrue($"Initial state of {nameof(BindingProcessImpl.InternalState.BindingOperationSucceeded)} should be true");
         }
 
-        [TestMethod]
-        public void SaveRuleConfiguration_CallsSolutionBindingOperation()
-        {
-            var solutionBindingOperation = new Mock<ISolutionBindingOperation>();
-
-            var testSubject = CreateTestSubject(solutionBindingOperation: solutionBindingOperation.Object);
-
-            testSubject.SaveRuleConfiguration(CancellationToken.None);
-
-            solutionBindingOperation.Verify(x => x.SaveRuleConfiguration(testSubject.InternalState.BindingConfigs.Values, It.IsAny<CancellationToken>()),
-                Times.Once());
-        }
-
         #endregion Tests
 
         #region Helpers
 
         private BindingProcessImpl CreateTestSubject(BindCommandArgs bindingArgs = null,
-            IBindingConfigProvider bindingConfigProvider = null,
             IExclusionSettingsStorage exclusionSettingsStorage = null,
-            IConfigurationPersister configurationPersister = null,
             ISonarQubeService sonarQubeService = null,
-            ILogger logger = null,
-            IEnumerable<Language> languagesToBind = null,
-            ISolutionBindingOperation solutionBindingOperation = null)
+            IQualityProfileDownloader qpUpdater = null,
+            ILogger logger = null)
         {
             bindingArgs = bindingArgs ?? CreateBindCommandArgs();
-            bindingConfigProvider ??= new Mock<IBindingConfigProvider>().Object;
             sonarQubeService ??= Mock.Of<ISonarQubeService>();
-            configurationPersister ??= Mock.Of<IConfigurationPersister>();
             exclusionSettingsStorage ??= Mock.Of<IExclusionSettingsStorage>();
+            qpUpdater ??= Mock.Of<IQualityProfileDownloader>();
             logger ??= new TestLogger(logToConsole: true);
-            languagesToBind ??= Language.KnownLanguages;
-            solutionBindingOperation ??= Mock.Of<ISolutionBindingOperation>();
 
-            return new BindingProcessImpl(bindingArgs, solutionBindingOperation, bindingConfigProvider,
-                exclusionSettingsStorage, configurationPersister, sonarQubeService, logger, false, languagesToBind);
+            return new BindingProcessImpl(bindingArgs,
+                exclusionSettingsStorage,
+                sonarQubeService,
+                qpUpdater,
+                logger,
+                false);
         }
 
         private BindCommandArgs CreateBindCommandArgs(string projectKey = "key", string projectName = "name", ConnectionInformation connection = null)
@@ -396,29 +166,6 @@ namespace SonarLint.VisualStudio.ConnectedMode.UnitTests.Binding
             connection = connection ?? new ConnectionInformation(new Uri("http://connected"));
             return new BindCommandArgs(projectKey, projectName, connection);
         }
-
-        private SonarQubeQualityProfile ConfigureQualityProfile(Mock<ISonarQubeService> sonarQubeService, Language language, string profileName)
-        {
-            var profile = new SonarQubeQualityProfile("", profileName, "", false, DateTime.Now);
-            sonarQubeService
-                .Setup(x => x.GetQualityProfileAsync(It.IsAny<string>(), It.IsAny<string>(), language.ServerLanguage, It.IsAny<CancellationToken>()))
-                .ReturnsAsync(profile);
-
-            return profile;
-        }
-
-        private class DummyConfigPersister : IConfigurationPersister
-        {
-            public BoundSonarQubeProject SavedProject { get; private set; }
-
-            BindingConfiguration IConfigurationPersister.Persist(BoundSonarQubeProject project)
-            {
-                SavedProject = project;
-                return new BindingConfiguration(new BoundSonarQubeProject(), SonarLintMode.Connected, "c:\\any");
-            }
-        }
-
-
 
         #endregion Helpers
     }
